Add remark about reading SharedValue from the js thread (#6160)
## Summary

Reading from a shared value on the js thread may block the JS thread
until the UI runtime finishes whatever it's doing. This isn't really
communicated clearly and adding it to the docs seems like the best
option as we don't want to show warnings.

## Test plan

I asked ChatGPT to point out problems (it did).