Tests of `withDelay` (#6150)
## Summary

Tests of `withDelay` consist may look a bit complicated, because they
contain a bit more logic than other tests. It's because I want to make
very peculiar snapshot tests that verify, if `withDelay` modifier really
keeps the same animation.
So I compare snapshots of animations with and withoutDelay, and they
should be the same (except for some static animation at the beginning).

This way it should be very quick easy to add more tests if we spot and
fix some specific bug related to this animation modifier.

One test case is commented out, as it doesn't pass - we have more than
one frame of unexpeced delay when using withDelay animation modifier.

## Test plan