docs: Add markdown header to docs (#6200)
Few docs were missing markdown header - instead they had docusaurus
title property. Because of that it wasn't possible to generate og images
for them. In the future these docs will be rewritten to meet current
style guide.

That's why we decided to add `# Title` in docs instead of fix for
`title: ...` in generate-og-images.js script. If we were to add fix to
the script it will become redundant within a month or two.

Following docs were added a header:
- interpolateColor
- Debugging worklets
- Testing with Jest