docs: Fix linking to `Contributing to the code` section (#6251)
## Summary

In the
[Contributing](https://docs.swmansion.com/react-native-reanimated/docs/guides/contributing)
section, right now in the 4th point of Ways to contribute subsection,
paragraph "contributing to code here" does not redirect to the given
subsection at all. This PR fixes this by changing the id of the
subsection to the correct one.
I've also changed `discord` to `Discord` in one of paragraphs, since
that small `d` has caught my attention 😄.