Put back the README.md file in the npm package (#6256)
## Summary

This PR adds the `README.md` file back to the NPM package. After
migration to a monorepo, the `README.md` file's location was changed to
the parent directory. Unfortunately, you can't include a file from the
parent directory in an npm package, and symlinks don't work for this
either.

## Test plan

run `npm pack` and check if README.md is inside