Rename ReJest main folder (#6292)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->
## Summary
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

Rename `ReanimatedRuntimeTestsRunner` into `ReJest`, because
RuntimeTestsRunner is not the only class exported from that folder, it
contains all the functionalities of reJest micro-framework.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->