Fix `Remove` mutation parsing in Layout Animations (#6286)
## Summary

This PR changes the way Remove mutations are parsed in
`LayoutAnimationsProxy::pullTransaction`. Before there was an assumption
that parent removals will be later on the list than removals of their
children. This is not accurate when a view is reparented. This
assumption meant that we could recognize "root" nodes (basically entry
point for exiting animations) while processing them. Now this part of
logic has to be calculated after all Remove mutations are parsed

fixes #6222

## Test plan
Check `[LA]` examples for regresions