Remove fontawesome dependency from example app (#6184)
## Summary
I noticed that `@fontawesome` dependency is present in the repository
but is used in 2 demo screens only - the size of unpacked dependencies
are following:
- `@fortawesome/fontawesome-svg-core`: 397 kB
- `@fortawesome/free-solid-svg-icons`: 5.09 MB
- `@fortawesome/react-native-fontawesome`: 33.4 kB

I believe that plain text emojis (which are present on eg. examples
list) could be used instead of fontawesome icons.

## Test plan
Open `Tab bar` and `Habits` example from the example app