docs: add disclaimer about ready-to-use packages in Reanimated Cookbook (#6182)
Reanimated Cookbook has a pure didactical purpose. While we provide the
full code some of the developers might think that they have to
reimplement the components every time.

In the ecosystem there are ready-to-use convenient packages that the
developers could use out of the box to solve their problems

This PR adds two recommendations about the Bottom Sheet and Marquee
example.