[reJest] Migrate to use the new runtimetests API (#6356)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Replace all uses of `FAILING` and `WARN` test decorators with
`expect().toThrow()`
The full removal of the logic behind these decorators is located here:
https://github.com/software-mansion/react-native-reanimated/pull/6358
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->