Fix github workflows (#6412)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

1. **Fix paths**
- Workflow files have been renamed, but the paths inside the workflows
weren't updated
- Due to this mismatch the workflow is not triggered once it is edited
4. **Fix caching**
   - Remove caching when building apps (it wasn't working well)
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
There are failing tests, but their fail in not depending on this PR

![image](https://github.com/user-attachments/assets/e30cef02-9175-4ca4-9442-4ab2115e0a98)

![image](https://github.com/user-attachments/assets/55730ecd-e691-49be-be80-194fd7ba5706)


<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>