docs: Fix incorrect links in docs (#6436)
## Summary

Some links were missing the `/` at the beginning, thus, they were
appended to the current page link instead of replacing it. As a result,
the **Page Not Found** page was displayed instead of navigating to the
desired page in docs.

![Screenshot 2024-08-21 at 11 18
51](https://github.com/user-attachments/assets/071b5e3a-7762-482b-95b1-a539363ab1da)

## Test plan

Just open reanimated docs and press on links that were changed in this
PR. You will see that before these changes, links do not navigate to the
desired page in docs.