Monorepo Support: Use node resolver to locate React Native package (#6482)
## Summary

Monorepos are configured and structured on what works best for the
project. Since this can vary from project to project, having a dynamic
way to find dependencies will make this work for more projects.
Currently, the repo only supports certain configurations of Monorepo
structures. This PR adopts a similar approach as Expo to locate the
react native package within the project. [Expo usage for reference](
https://github.com/expo/expo/blob/9821f2f8c8927d930a15041f22cd55e28f443274/packages/expo-modules-core/README.md?plain=1#L65)

Fixes #5297 and those related to it.

This PR increases overall support for monorepos, but does not change the
functionality of the project. It maintains the support for the
`REACT_NATIVE_NODE_MODULES_DIR` environment variable.

## Test plan

1. `yarn && yarn build`
2. `cd apps/fabric-example && yarn start`
3. `yarn android` <- Builds Fabric Example App for Android
4. Successfully builds Android App
5. Repeat steps 2-4 for the `apps/paper-example`

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>