Fix issues with skipping reanimated commit (#6498)
## Summary

When working on #6330 I missed the fact that the
`pleaseSkipReanimatedCommit` mechanism is not only a performance
optimization (preventing us from applying some changes twice a frame).
It's main role is to actually prevent reanimated from starving RN out.
It could happen that RN is trying to apply some heavy commit while an
animation is running. If during that commit reanimated applies its own
changes, then the RN commit will fail and another attempt will be made.
Since reanimated runs a commit in every frame, then the RN commit might
continue failing until the animation is over, which is very very very
bad.

In this PR I changed the "skip" name to "pause" since it better
describes the intent behind it. I also reverted the change where the
skip flag was reset in `performOperations`. This was done to ensure that
if some changes are applied through `updateProps` after the commit hook
has already gathered all updates, but before reanimated commits have
been unpaused, then these changes are not lost. Unfortunately this would
lead to reanimated blocking heavy RN commits from being applied. Now the
skip flag is reset in the `ReanimatedMountHook` (as it was before
#6330), and if there are any changes that would be lost by skipping
reanimated commits, we run an additional commit in the
`ReanimatedMountHook`

## Test plan
Check if the example app works fine