Update ReducedMotionConfig.mdx (#6598)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Adding warning that ReducedMotionConfig is applied globally, as I was
checking and docs and did not know until I checked the
[pr](https://github.com/software-mansion/react-native-reanimated/blob/0bd65b35bbe33c69590e4bfcb836f618ff066534/packages/react-native-reanimated/src/component/ReducedMotionConfig.tsx)
this component was introduced in

## Test plan

check the website after it gets deployed

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>