Fix crash on macos when using Layout Animations (#6600)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Previously we were passing a window to function convertPoint, that
accepts only view. It was causing a crash, therefore we should pass the
contentView of the window.
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
Tested on several examples:
* Olympic example
* [LA] Basic Layout Animation
* [LA] Entering and Exiting with Layout 
* [LA] Carousel
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->