Move headers to `apple` directory (#6612)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

This PR moves iOS/macOS/tvOS headers from `apple/reanimated/` to yet
another `apple` subdirectory so that the file tree reflects the paths in
`#import <...>`.

After this change, `apple/reanimated` can be added to
`HEADER_SEARCH_PATHS` and imports like
`reanimated/sensor/ReanimatedSensor.h` will work properly.

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->