Do not make a synchronous call when canceling animation.  (#6564)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
![Screenshot 2024-10-02 at 14 00
50](https://github.com/user-attachments/assets/12a0a1c4-bfc1-498f-950a-68f4da49dc9a)
I noticed in client's code that cancelAnimation called for few values
takes almost a second.
That seems to be caused by the fact that it's used on every
useShareValue hook on unmount and across multiple libraries build on the
top of reanimated. Looks like getter is synchronous and blocks the js
until we schedule a block on UI.
I suspect the idea was to make sure we use the same type as we can
animate different types. I thought it can be possible to just take
initial value if it's set. Please let me know if for some reason we have
to make a synchronous call either way.
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->


## Test plan
- Start animation and then cancel it with cancelAnimation function
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>
Co-authored-by: Tomasz Żelawski <tomasz.zelawski@swmansion.com>