chore: Bump web example to 0.76 (#6679)
## Summary

When working on
https://github.com/software-mansion/react-native-reanimated/pull/6622 I
noticed the web example app is not bundling. I found out it stopped
working when the rest of the examples were updated to RN 0.76.

This PR bumps the version of `react-native` and `expo` fixing the
problem and making it compatible with the rest of the repository.

## Test plan

Build and run the web example