Update ReanimatedPackage to supported stable API (#6730)
## Summary

Replace the deprecated `TurboReactPackage` with the newer
`BaseReactPackage` to improve compatibility with future React Native
releases. Note that this change requires at least react-native@0.74+,
but this should not be an issue as, according to the compatibility
table, Reanimated@3.17.0 only supports versions 0.74 and above.

Similar change has been introduced in the SVG package
https://github.com/software-mansion/react-native-svg/pull/2541

## Test plan

CI should pass