Fix bug in useSharedValueMock (#6742)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

The API for sharedValue has changed from using .value to get() and set()
methods. As a result, the mock that previously returned an object with a
single value property no longer works in the tests.

## Test plan

This is a PR for the mocks, so it's not necessary for it.

---------

Co-authored-by: Tomasz Żelawski <tomasz.zelawski@swmansion.com>