fix: don't clobber keyboard height when keyboard did not animate on iOS
When processing a keyboard frame, if the keyboard did not animate to the
target height (e.g. when closing the software keyboard on a device with
a physical keyboard), the height would be set to the wrong value in
updateKeyboardFrame on the subsequent update.

Instead of relying on the measuringView, which refers to the closing
animation in such cases, return _targetKeyboardHeight, which is also
returned in keyboardWillChangeFrame.