fix: restore `classic` runtime for @babel/preset-react (#6760)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

Fixes
https://github.com/software-mansion/react-native-reanimated/issues/6665

Applying the change I made to `react-native-builder-bob`:
- https://github.com/callstack/react-native-builder-bob/pull/695

It restores `classic` runtime for the `@babel/react-native` preset used
for generating ESModule code. The behavior was changed when we bumped
builder-bob:
- https://github.com/software-mansion/react-native-reanimated/pull/6485

I ran the `diff` for the version of `lib` without `jsxRuntime: classic`
and with it - the differences are only in regards to `_jsx` usage. You
can read more here:
-
https://github.com/software-mansion/react-native-reanimated/issues/6665