fix: don't prematurely update keyboard height during dismissal on iOS
When the keyboard is dismissed via the close button or is undocked, the
interactive dismissal observer will be notified of the geometry change
before the show/hide notification is received.

Bail out before updating the keyboard height if the reported keyboard
height is 0, if the keyboard state is CLOSED or if the keyboard is
floating (keyboard width != window width.)

Note that this still allows some updates through when they shouldn't be,
but it's better than before, especially for floating keyboards.