refactor: restore JSCallInvoker (#6783)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Turns out that majority of work with regards to using `RuntimeExecutor`
done in
- https://github.com/software-mansion/react-native-reanimated/pull/5734
can be reverted since `jsCallInvoker` is staying in the API.

Thanks to that we can significantly cleanup the code which is helpful
for Worklets.


## Test plan

I tested the changes on 0.74, 0.75 and 0.76 for Android/iOS Paper/Fabric
and all worked well.