refactor(Worklets): move JSScheduler (#6762)
## Summary

This pull requests moves `JSScheduler`, a class responsible for
scheduling operations back to the JavaScript thread, to Worklets.

- `JSScheduler` is now available from `WorkletsModuleProxy` in cpp and
that's how `ReanimatedModuleProxy` obtains it when it needs it. More
sophisticated APIs will be made in the future.
- Both `ReanimatedModule` and `WorkletsModule` are now responsible to
forward `jsCallInvoker` to their own Proxy modules. `jsCallInvoker` is
no longer available in `JSScheduler`.

## Test plan

CI