fix: correct height when keyboard opened from shortcuts bar on iOS
UIKeyboardDidShowNotification on iOS 18.2 returns an incorrect end frame
origin when opening the on-screen keyboard from the shortcuts bar when a
hardware keyboard is attached.

Instead of relying on this incorrect frame, return the height of the
keyboard view in response to UIKeyboardDidShowNotification and 0 in
response to UIKeyboardDidHideNotification.

State changes are now completely handled in response to notifications,
and updateKeyboardFrame is only responsible for animation tracking.