fix: __reanimatedLoggerConfig is not defined on edge runtime (#6880)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

If you run NextJs application in edge runtime, library throws an error

`__reanimatedLoggerConfig is not defined`

we needs to declare global or globalThis on every usage in workers on
edge runtime.


## Collaboration
**It was found and made in collaboration with @vault-developer** 


## Test plan

Tested with patch-package and it worked