fix: return correct height for detached keyboard on iOS
Previously, when the keyboard was detached, useAnimatedKeyboard returned
a height corresponding to the height of a docked keyboard. However, when
the keyboard is floating, undocked or split, the height should be 0.

Listen for the UIKeyboardWillHideNotification notification to determine
if the keyboard has been detached. We must now listen for show and hide
notifications instead of UIKeyboardWillChangeFrameNotification, as the
show/hide notifications are dispatched after the change frame
notification.

Adapted from the initial work in
https://github.com/software-mansion/react-native-reanimated/pull/5710.
Listening for hide is simpler than calculating whether the keyboard is
detached, compatible with Stage Manager, and maintains compatibility
with interactive keyboard dismissal.