fix: Break circular reference in `LayoutAnimation` (#6671)
Breaks circular reference in the `LayoutAnimations`.

`LayoutAnimations ` is a `HybridClass` with a C++ counterpart holding a
global reference to the Java object. This structure creates a circular
reference between C++ and Java, which the garbage collector cannot clean
up. To resolve this issue, I manually removed the reference between C++
and Java by resetting the global ref during the destructor of the
`NativeProxy`.

- tested in Expo Go ✅

---------

Co-authored-by: Tomasz Żelawski <tzelawski@gmail.com>