fix: Cocoapods versions used in macos and tvos examples (#6926)
## Summary

`Gemfile.lock` files in `macos-example` and `tvos-example` contained
`cocoapods (1.16.1)` in spec. Because of that, even though the
`Podfile.lock` had `cocoapods (1.14.3)` specified at the end, a newer
version of cocoapods was used to install pods and the cocoapods version
in the `Podfile.lock` was always updated to the new one. I reverted them
back to `1.14.3` to match other example apps.