chore: Clean up dependencies of apps dependent on the common-app (#6923)
## Summary

This PR moves common-app dependencies to the common-app `package.json`
and removes all of duplicate dependencies from `fabric-example`,
`web-example` and `macos-example` apps. We can do this thanks to
`react-native.config.js` which reads all dependencies from the
`common-app/package.json` and adds them to the dependency list from the
current app's `package.json` file (if they weren't specified in the
app's `package.json`).

Some dependencies, which are used in `package.json` scripts to e.g.
start the app, such as `react-native`, must be added explicitly in each
app's `package.json`.

### Motivation

I think it is a bit cumbersome to repeat dependencies needed by the
common-app in every app that imports application from the common-app
sources. Since every app would need dependencies that common-app needs
to work, we don't have to repeat them in every app dependent on the
common-app.

If we will have to use a different dependency version or mock a
dependency, we can always add another dep e.g. in the fabric-example
`package.json` or exclude/include it in the `react-native.config.js`.

## Test plan

Build the `fabric-example` / `web-example` / `macos-example` to see that
everything still works.