Fix layout animation crashes on the New Architecture (#6920)
## Summary

This PR adds a `contains` check for layout animation config. When we
schedule the start of an animation, we check if an animation is
configured for the given view. But it might happen that before the
`LayoutAnimationsManager::startLayoutAnimation` function gets executed,
the view gets removed. When the view is removed we clear the config, so
the `startLayoutAnimation` function will get a `nullptr` config,
resulting in a crash.

fixes #6908

## Test plan