chore: Bump RN version to 0.77 stable (#6933)
## Summary

This PR bumps version of RN to `0.77.0` stable. In most cases it is just
a change from `0.77.0-rc.6`, but in the `tvos-example` I bumped version
from `0.76.0` to `0.77.0`.

I also fixed tvos-example on android (it seems that nobody have ever
built it on android as paths to `node_modules` in `settings.gradle` were
invalid).

## Tested example apps

- [x] fabric-example iOS
- [x] fabric-example Android
- [x] web-example
- [x] macos-example (I left RN 0.75.4 unchanged until we will be
upgrading to Fabric)
- [x] tvos-example (iOS)
- [x] tvos-example (Android)

## Recording of working tvos example


https://github.com/user-attachments/assets/9a753a96-154f-4820-961d-6cd1b6f11c0d