Introduce color interpolation based on the LAB color-space (#6782)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

In this PR, I've introduced color interpolation based on the
[OKLAB](https://en.wikipedia.org/wiki/Oklab_color_space) color space.

<img src="https://facelessuser.github.io/coloraide/images/oklab-3d.png"
width="33%" height="33%"/>
<sup>(image credit: <a
href="https://facelessuser.github.io/coloraide">ColorAide</a>)</sup>

### Details

OKLAB is derived from CIELAB - both considered improvements compared to
RGB and HSV/L in terms of color gradient computation, as the linear
interpolations over their channels results in a smoother experience.
That's for 2 reasons:
1. The color space tends to better preserve light intensity across the
transition (largely thanks to the L channel)
1. They minimize the inter-color transitivity

**L\*a\*b\* does often yield results very similar to RGB w/ gamma
correction, as in the case of the default colors on the web-site:**

![image](https://github.com/user-attachments/assets/e12aefcc-e9c9-49c3-87b1-79b9574da694)

Nevertheless, there are some cases where the result can be considered
better. Going from red to blue, for example, LAB shows a smoother
transition with less intermediate coloring (purple in this case):

![image](https://github.com/user-attachments/assets/30b7ea67-d185-4245-b0f0-513b20d9ad1b)
Also, in yellow-to-blue, LAB better preserves lightness:

![image](https://github.com/user-attachments/assets/8db0b723-4ff2-4c12-8ffd-4ad631834e5a)


### Things to notice

* In my implementation, I've completely relied on `culori`, but had to
rip their code off from it (all-the-while keeping the credit). I
couldn't use it as-is, because I had no way to specify it's meant for
off-loaded execution (i.e. using the `worklet` notation). **I could
definitely use some advice with that** by the maintainers ❗
* In terms of user-facing API's, I've decided to name the color-space
selection `LAB` (alongisde `RGB` and `HSV`), hiding the implementation
details (i.e. the underlying selection of _oklab_, in particular). I
believe that any future offering of other LAB spaces (e.g. CIELAB) could
be introduced by allowing for LAB-specific `options`.
* OKLAB is the default implementation in native Android development,
offered by Jetpack Compose.

## Motivation

There are potential screens in the apps we're working on where a smooth
color transition could improve the user experience. We're avid users of
`reanimated` and would love to be able to incorporate the best
experience though it.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

TBD