Add Keyframe animation caching (#6960)
## Summary

Our current implementation of keyframe parsing mutates the keyframe
definition object (e.g. we remove the easing function). If we try to
call `build()` on a `Keyframe` object more than once, we will end up
with a different animation after the first time. This is bad because we
want people to memoize their Keyframes. I decided to fix this problem by
caching the calculated animation. This fixes this problem, and also
allows us to reuse the result of parsing.

## Test plan