feat(Worklets): JS code duplication (#6921)
## Summary

In this PR I'm moving the `WorkletsModule` part of the JS code to
`react-native-worklets`. Reanimated would still have those files
duplicated via a mechanism introduced in #6827.
Now in Reanimated all worklets code must be acquired via a special
`WorkletsResolver` module. This is a temporary solution and will be
dropped once we require `react-native-worklets` as a mandatory
dependency for `react-native-reanimated`.

- Added a workaround so that in our monorepo `react-native-worklets`
wouldn't be used in the app even though it's not a dependency.
- Added an ESLint rule that enforces imports from `WorkletsResolver`.
- I duplicated `PlatformChecker.ts` explicitly in
`react-native-worklets`. This file is small and stable so I don't
consider this to be an issue. We can think of some better solution once
the package is more polished.
- Split some types from `commonTypes.ts` between packages.


## Test plan

Fabric example should work with & without `react-native-worklets`
installed. You can check using the debugger that always the correct
instance of Worklets' JavaScript code is used.