refactor: move Worklets JS core from Reanimated (#6973)
## Summary

Currently JS core of Worklets sits in Reanimated package. This PR moves
it to Worklets. It's a bit messy but we'll get there eventually.

- Added `createCustomError` that creates `WorkletsError` and
`ReanimatedError` used in respective packages.
- Added ESLint rule `use-worklets-error`, a copypaste of
`use-reanimated-error`.
- Moved `shareables.ts` to `react-native-worklets`.
- Moved `threads.ts` to `react-native-worklets`.
- Moved `initializers.ts` to `react-native-worklets`
- Moved `runtimes.ts` to `react-native-worklets`
- JS version checking in worklets is disabled for the time being, as it
would clutter this PR even more. We'll circle back to it once we set
version checking between Reanimated and Worklets.

## Test plan

CI, run the example app.