fix: move requestAnimationFrame back to Reanimated (#6991)
## Summary

I introduced a regression in #6973 by moving JS implementation of
`requestAnimationFrame` to Worklets. Native version of
`requestAnimationFrame` is injected in Reanimated's
`UIRuntimeDecorator.cpp` so `requestAnimationFrame` must be overwritten
in Reanimated, not in Worklets.

Since all the native implementation is in Reanimated and it would be a
hard task to extract it right now, I decided to keep it for the time
being unless we decide where it should land eventually.

## Test plan

The only difference is the lack of batching with my regression
introduced. You can `console.log(currentCallbacks.length)` to see that
the batching is restored.