Remove UI runtime usage by CSS Animations (#7006)
## Summary

The goal of Reanimated 4 is to enable the use of CSS Animations without
needing access to the UI runtime. This PR removes references to the UI
runtime within the CSS codebase. It's important to note that in some
cases, we still have access to a runtime, but this is a React Runtime
used on the React thread.

There is no performance degradation after the rewrite to
`folly::dynamic`. It's even a bit faster, but the improvement is within
the measurement error range.

---------

Co-authored-by: Mateusz Łopaciński <lop.mateusz.2001@gmail.com>