chore: Make CSS interpolators stateless (#6993)
## Summary

This PR removes, probably unnecessary, premature optimizations which
added code complexity to interpolators implementation. Interpolators
kept state of the current keyframe and last interpolated value. They
also were tightly coupled with progress providers which must have been
passed in the constructor.

Even though the previous implementation worked fine, it was impossible
to optimize creation of the animation (we had to create interpolators
for all views separately, even if they used the same animation). This
also resulted in higher memory consumption.

Instead of moving left or right on the keyframes sequence, I now use a
binary search to find the current keyframe. It may have worse time
complexity (`O(log(n))` instead of `O(1)`) but, in fact, may be even
faster because of removed logic responsible for storing a previous
value, iterating over keyframes, etc. This change shouldn't affect
performance as long as somebody doesn't use thousands or keyframes in
animations (which is rather impossible).

## How it was tested?

I just went through examples in the example app (on both, iOS and
Android) to make sure that no regression was introduced in this PR.