docs: Add missing information about animated styles precedence (#7070)
## Summary

Refers to the issue: #7067

This PR adds missing information about animated styles behavior to the
remarks section of the `useAnimatedStyle` documentation page.

We should add proper precedence support in the future, but for now,
until we figure out a valid solution to this problem, we should include
information about this behavior in docs.

## Screenshot

![Screenshot 2025-02-21 at 18 20
23](https://github.com/user-attachments/assets/7feed974-9ac3-406b-ba52-0eb224713b33)

## Test plan

Build docs and open the `useAnimatedStyle` hook docs.