Add extraPlugins and extraOptions to babel plugin (#7072)
## Summary

In React Native projects that use Flow using Flow types inside worklets
crashes the app when transforming the code.
Here is a repo that reproduces the crash:
https://github.com/graszka22/reanimated-flow-repro.
A solution is to add `@babel/preset-flow` to reanimated plugin. Because
Flow is now rarely used and users might want to add other plugins, I've
added options: `extraPlugins` and `extraPresets` so anyone can configure
it however they want.
## Test plan

Used the code from the repro
https://github.com/graszka22/reanimated-flow-repro to test. Just verify
that the app is no longer crashing.

---------

Co-authored-by: Tomasz Żelawski <tzelawski@gmail.com>