Merge `NativeProxyCommon` into `NativeProxy` (#7122)
## Summary

We used to have more than one version of `NativeProxy` along with
`NativeProxyCommon` to minimize code duplication. Once we dropped
support for Paper in
https://github.com/software-mansion/react-native-reanimated/pull/7106,
now there's only one `NativeProxy` so it makes sense to merge it back
with `NativeProxyCommon`. This PR does exactly that.

Suggestion: You can review the PR commit by commit.

## Test plan