Add asserts for queue during Worklets initialization on iOS (#7137)
## Summary

This PR adds `RCTAssertMainQueue()` and `REAAssertJavaScriptQueue()`
calls in various initialization methods in `WorkletsModule`,
`AnimationFrameQueue` and `WorkletsMessageThread` on iOS along with the
implementation of `AssertJavaScriptQueue`. This way we can make sure
that the methods are called on appropriate threads as well as track this
information in the codebase.

I wasn't able to come up with an assert for
"com.meta.react.turbomodulemanager.queue" in `invalidate` methods since
`NSThread.currentThread.name` returns `nil` so I just added `// Called
on com.meta.react.turbomodulemanager.queue` comments.

## Test plan

✅ Tested in debug and release mode