Remove `requiresMainQueueSetup` with `YES` (#7133)
## Summary

⚠️ This is potentially a very breaking change in some setups!

This PR removes `requiresMainQueueSetup` method in `REAModule` that was
responsible for marking that Reanimated module should be initialized on
the main queue.

This method was added as a side effect 3 years ago in
https://github.com/software-mansion/react-native-reanimated/pull/3555 in
order to make the warning go away and to keep the identical
initialization behavior.

Looks like this is a perfect moment to get rid of this chunk of the code
and hopefully it should work just fine.

Please let us know immediately if you have any troubles with running
Reanimated without `requiresMainQueueSetup`.

## Test plan