Remove `methodQueue` with `RCTGetUIManagerQueue` (#7156)
## Summary

⚠️ This is potentially a very breaking change in some setups!

This PR removes `methodQueue` method in `REAModule` that was responsible
for marking that native methods of Reanimated module should be
initialized on the UIManager queue.

As stated in the comment, `REAModule` methods needed to be executed on
the same queue as the `UIManager` to avoid having to lock `_operations`
and `_preOperations` since `uiManagerWillPerformMounting` will be called
from that queue. Since `_operations` and `_preOperations` have been
removed, we can finally get rid of `methodQueue` as well.

After this change, invalidation methods will be called on the
TurboModuleManager queue instead of the UIManager queue.

Looks like this is a perfect moment to get rid of this chunk of the code
and hopefully it should work just fine.

Please let us know immediately if you have any troubles with running
Reanimated without `methodQueue`.

## Test plan