refactor: run `_maybeFlushUIUpdatesQueue` through `_microtaskQueueFinalizers` (#7064)
## Summary

With current implementation of `queueMicrotask` we call
`_maybeFlushUIUpdatesQueue` which executes `performOperations`. This
isn't a good pattern in my opinion, since `queueMicrotask` shouldn't be
tied directly to the rendering layer. Regardless, to keep the current
behavior I added `_microtaskQueueFinalizers` array in
`react-native-worklets` where Reanimated could register its finalizer -
it has to be executed after all the tasks, to gather all the updates.

This should be considered as a temporary API and we'd need to overhaul
this in the future for a consistent event loop model.

## Test plan

See that animations run just fine!