fix: Crash caused by passing animated ref to the animated component (#7158)
## Summary

This PR fixes crash caused by invalid call of `hasAnimatedRef` method.

I don't know why methods created in this way:
```tsx
hasAnimatedRef() {
  return ...
}
```
don't work whilst method assigned as props of the object work:
```tsx
hasAnimatedRef = () => {
  return ...
}
```

You can see the difference when you log objects with both methods. In
the first case, there is no `hasAnimatedRef` method in the object.

I decided to get rid of `hasAnimatedRef` method and use the
`_hasAnimatedRef` property because we already access the `_componentRef`
prop directly, so we can do the same with the `_hasAnimatedRef` prop
instead of adding a function that returns this value.

## Test plan

Copy the example from the linked issue or from below (it is the same)
and paste it to the `EmptyExample`. You will see the difference.

<details>
 <summary>Example code</summary>

```tsx
class ClassComponent extends Component {
  render() {
    return <View />;
  }
}

const AnimatedClassComponent =
  Reanimated.createAnimatedComponent(ClassComponent);

function App(): React.JSX.Element {
  const ref = useAnimatedRef();

  return (
    <SafeAreaView style={styles.container}>
      <AnimatedClassComponent ref={ref} />
    </SafeAreaView>
  );
}
```
</details>