Remove implementation of `UIManagerListener` and `UIManagerModuleListener` interfaces in `ReanimatedModule` on Android (#7169)
## Summary

Currently, `mOperations` in `ReanimatedModule` is never populated, so it
can be removed together with `UIThreadOperation` interface and
`willDispatchViewUpdates` interface methods from
`UIManagerModuleListener` and `UIManagerListener` interfaces, along with
all the remaining methods which are currently no-ops. This means we also
need to remove `addUIManagerEventListener` and
`removeUIManagerEventListener`. Turns out, we don't need
`FabricUIManager` as well. After all these changes there will be only
one runnable to execute in `invalidate` method so I also simplified
`mUnsubscribe` assignment.

## Test plan

Build, launch and play with fabric-example on Android.