Remove `Utils.isChromeDebugger` on Android (#7187)
## Summary

Chrome Debugger is now deprecated and will be removed in RN 0.79 so we
don't want to support it in Reanimated 4.

This PR removes Chrome Debugger-related checks in Android codebase (we
already don't have any checks on iOS).

I assume this is still a safe change in terms of error reporting since
we still have `isChromeDebugger()` check in TypeScript.

## Test plan

Build and launch fabric-example on Android.