feat(worklets): prepare releasing pipeline (#7107)
## Summary

To start releasing nightlies of `react-native-worklets` and make the
process of releasing it similar to how it's done for
`react-native-reanimated` we need to refactor some bits.

I extracted the core of determining the version (custom, nightly or
fresh) to a separate script in the root of the repo.

## Test plan

Run the scripts and see that the packages are built correctly.