refactor: remove runOnUIImmediately (#7226)
## Summary

`runOnUIImmediately` is not conforming with a well defined event loop
model of the UI Runtime, so we should remove it.

I considered that we can replace it with:
1. `runOnUI`
2. `executeOnUIRuntimeSync`
3. Additional batching mechanism on the side of Reanimated, using
`runOnUI`.

Of those I went for **1.** for the sake of simplicity and didn't notice
any regressions. **2.** would be the second choice since it actually
preserves the logic but could be potentially slowing down the
initialization process.

Please submit any suggestions on how to test this change more
thoroughly.

## Test plan

Run an example and see that all animations run well 🚀