chore: Add css namespace for CSS-related code (#7213)
## Summary

This PR adds a `css` namespace under the `reanimated` namespace for
CSS-related code. It also removes the redundant `Interpolators`
namespace.

I decided to add this namespace because there is much CSS-related code
(and will be more) so it would be safer to keep this code separate from
the remaining reanimated code. Since, CSS is a standalone feature
independent on other features, it seems to be reasonable to keep the
respective code in a separate namespace.