fix: CSS issues introduced by react-freeze fix (#7237)
Opened this PR as a replacement for #7229

It fixes issues introduced in the #7053 PR and keeps the necessary logic
to ensure that `react-freeze` fixes work as well.

Basically, the problem is that we cannot trust `componentWillUnmount`
calls because they are also called when component gets frozen (we don't
want to remove animations in this case - it will be good to pause them
in the future, but the original PR with react-native fix also didn't do
that, so it will be addressed in the future).

To make it work, I decided to use `componentWillUnmount` only to mark
nodes as removable (possible to remove). The final decision is made in
the reanimated mount hook, where we check if the node was removed and
clean up registries only if it no longer exits in the ShadowTree. In all
other cases, we keep the node marked for removal as long as it is frozen
and unmark it when it gets unfrozen.

1. Open the app, navigate between screens, see that everything works.
2. Open the drawer navigator and the *React Freeze* example to observe
that everything behaves as expected.
3. Navigate between screens in this example and press on the *Check for
registry leaks* button to observe that there are no leaks.