fix: Exclude deprecated, unsupported properties from style type (#7300)
## Summary

This PR excludes deprecated props that aren't supported by CSS
animations (and shouldn't be used in animated styles as well, as they
are deprecated) from the `PlainStyle` type.

The issue was originally reported by @softwarebyze and was related to
CSS transitions allowing to e.g. pass the `translateX` string for the
`transitionProperty`, which is not supported and `transition` string
should be used instead.

### Before

![Screenshot 2025-03-25 at 14 14
41](https://github.com/user-attachments/assets/e3be2461-66ad-4780-9d14-80b32c801d19)

### After

![Screenshot 2025-03-25 at 14 13
28](https://github.com/user-attachments/assets/d4ea7c19-d620-480f-b1d4-7dfd418c2c0b)