Fix react-freeze (#7053)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

This PR is very annoying. Until now we assumed that when
`componentWillUnmount` is called, the component will never reappear.
This is false when using `<Suspense>`
([docs](https://react.dev/reference/react/Suspense#caveats)), which is
exactly what `react-freeze` does. To fix the issue I had to move our
animation cleanup functions to be called only when we are sure that the
component will not reappear. This is how the new approach works for each
registry:
- `AnimatedPropsRegistry`: we remove a record from this registry only
when we fail to apply it onto the ShadowTree (in the commit hook or
`perfromOperations`)
- `CSSAnimationsRegistry`: we remove a record from this registry only
when we fail to apply it onto the ShadowTree (in the commit hook or
`perfromOperations`)
- `CSSTransitionsRegistry`: we remove a record from this registry when
`componentWillUnmount` is called and there is no ongoing transition OR
we failed to apply it onto the ShadowTree (in the commit hook or
`perfromOperations`)
- `StaticPropsRegistry` remains the same as it doesn't store any state
that needs to be persisted when freeze is used

One problem with this approach is that we are running animations on
components that are freezed. This is a waste of resources. This will be
addressed in a separate PR, but the main idea is to pause animations in
`componentWillUnmount` and resume them (with the new timestamp) in
`compnentDidMount`.

Also currently the leak check will say that `AnimatedPropsRegistry`
leaked if we don't reload it. This is because an animation frame comes
after the cleanup from the commit hook. Normally this would be cleaned
up in `performOperations`, but because of the `sunchronouslyUpdateProps`
fast-path it won't. The cleanup will still happen on the next commit
hook invocation (i.e. if you hit reload on the leak check). This issue
soon will be gone as we are removing the fast path in #7014.

| Before | After |
| -------- | ------- |
| <video
src="https://github.com/user-attachments/assets/249acaac-f1f0-4846-a38e-4dab54cfb49b"/>
| <video
src="https://github.com/user-attachments/assets/5d3c578b-fb80-45e8-9955-542bf47d0573"/>
|

Go through the app, and then visit the `FreezeExample`. Use the `check
for registry leaks` button.