(chery-pick) feat: prevent flattening style arrays (#7327)
## Summary

Cherry pick of #7021. Main reason for it was that the early return of
`currentStyles` (if `style` is an array) caused to not apply
`jestAnimatedStyle`. Becasue of that execution of tests using `style`
arrays would fail.

I decided to cherry-pick whole PR not only desired fragments to keep
integrity between `main` and `3.17-stable`.

## Test plan

---------

Co-authored-by: Jacek Pudysz <jacekpudysz@gmail.com>