release: 3.17.2 (#7332)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Release of Reanimated 3.17.2. Cherry-picks can be found here #7207 

## Test plan

Tested in our example app (Fabric, Paper, tvOS - every on iOS and
Android, and MacOS) and in Clean App (RN and RN web), both prod and
debug.