Bump react-native-is-edge-to-edge to 1.1.7 (#7346)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

This PR unfixes the current `react-native-is-edge-to-edge` version.

The current version has `peerDependencies` that prevents the Expo team
from creating a canary release:

```
Conflicting peer dependency: react-native@0.78.2
```

npm installs this version to respect the `>=0.73.0` requirement, even if
`react-native@0.79.0-rc.0` is already a project dependency.

PS: I'm on my phone and without a computer access, if someone can run
yarn and fix the pipeline before merging, it would help 🙏

---------

Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>